backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
streamlit
pandas
numpy
scikit-learn
joblib
openai
fpdf2
passlib[bcrypt]
PyJWT
